nonce manager uint64

+++ src/core/NonceManager.sol
@@ -11,7 +11,7 @@ contract NonceManager is INonceManager {
     /**
      * The next valid sequence number for a given nonce key.
      */
-    mapping(address => mapping(uint192 => uint256)) public nonceSequenceNumber;
+    mapping(address => mapping(uint192 => uint64)) public nonceSequenceNumber;
 
     function getNonce(address sender, uint192 key)
     public view override returns (uint256 nonce) {
